feature: Create domain package and move domain class
